WIP Close container on destruction
I'm not sure of the implications of this, so putting it out there for
comments. This might help people who forget to call `Mockery::close`,
which happens often.

We could probably catch our exceptions and restyle in a "Did you
forget..." style. I'm sure someone did something similar to this
already, but I don't know where it ended up...